refactor(bigquery/storage/managedwriter): improve readability
This PR introduces no functional changes.  It simply colocates the
two types of variadic options into the same file to aid readability.

Towards: https://github.com/googleapis/google-cloud-go/issues/4366
